chatapp/requirements.txt
uvicorn[standard]
Django>=4.0
djangorestframework
channels
channels_redis
djangorestframework-simplejwt
